+++ examples/reflecting_laser.rs
@@ -10,7 +10,7 @@ fn main() {
     App::new()
         .add_plugins((
             DefaultPlugins.set(bevy_mod_raycast::low_latency_window_plugin()),
-            DefaultRaycastingPlugin,
+            CursorRayPlugin,
         ))
         .add_systems(Startup, setup_scene)
         .add_systems(Update, bouncing_raycast)